OK. Cleaned up things a bit. Provided code for BerkDB vers 2, 3, and 4. They key off the DB_VERSION_MAJOR from db.h. The makefile was changed to bump the version to 0.9, and include the -ldb and whatnot where necc.
On the local machine, all I've proved is the reverse listing doesn't crash,
that the package comes up, and seems to operate. The pid file gets written.
The db code compiles and might work. And even if it doesn't work, as long as it
doesn't crash, we are in business.

Up to this and the last version, behavior has been good. This filter really
takes a very significant chunk of spam out of your site's diet. I'm seeing
personally over 90% success, and with a few judicious blacklist entries, I've
reduced the spam by nearly 100% some days. Usually, some oddball shows up. And a
few that are kind of "borderline" spam.

One thing I have to work on is implementing the limiting of threads. There's
no hooks to do this in the milter interface. I get crashes every few days that
dump humongous cores. The best thing to do at the moment is a daily kill and
fresh start of the relaydelay process. That's what got me started.

Also, I'm not really done yet with including the reverse_track calls everywhere
in the code. More in future releases.

The extra config file variables have been added to the parser, and print out
in the startup.

And, one last concern. The perl version seems to have reorganized a little, with
code being shifted here and there. I'll have to review it all and make sure
this code follows suit.

I've also pretty much proved that the blacklisting and whitelisting works from
the mysql db. The "from" keyed blacklisting is very handy. "From" keyed whitelisting
has become very handy, as long as no local addresses are used, it can be used
where just ip's cannot. And, the "from - to" combo is great for whitelisting
people's address books. So, the spammer barbara "likes", isn't going to get anyone
else at the site.