Added code to limit length of Rcpt address to 254 chars. A Reject is issued if it longer. This is way too long. It smells of some kind of exploit. You can get really long, valid addresses by using the + comment notation via the SMTP spec. Example:
murf+areallylongandtortuousadditionalstringsurelymeanttostresssomeaspectofthemailingsystemandmakeitfailinsomegruesomeandpotentiallyusefulwaytosomehackerthatwouldcertainlyleadtonogood@e-tools.com

Also, I suspect that some mailers may have already adapted to the few of us greylisting, and
may hopefully be paying attention to the tempfail message. So, as an experiment, I have
invented 40+ different tempfail messages, and the milter randomly deals these out as
necessary. It might be best if users invent these themselves. Maybe we need to randomize
the result codes a little also...

murf