Integrated a patch from Martin A. Walker <maw@synesis.net> to allow logging to a logfile and running as a daemon.  Actually integrated this a long time ago, but forgot to check it in.
Changed the accessdb checks to use BerkeleyDB module instead of DB_File, as
it appears to be more compatible with the db versions sendmail uses on more
distributions.  Left lines in source but commented out for using DB_File.
In case anyone needs to, they can revert to using DB_File easily.

Added a check to verify that the dns_name table exists if name updates are
enabled.